Add vpp file for UMLS
this is the file to edit for the different UML diagrams. As per the assignment discretion 
"
One tool you can use for the UML diagrams (activity, class, and use case diagrams) is called Visual Paradigm. All the diagrams should all be part of the same conceptual model. With Visual Paradigm, this means all of the diagrams should be in one VPP file.
"